fix(refactor): Refactor Legrand specific FZ
* Moved legrand_cluster_fc01 to Legrand specific library
* Renamed it to cluster_fc01
* Updated references following refactoring
